Rename `piece_by_key` module to `piece_by_index` and remove unused request handlers
