+++ crates/subspace-networking/src/request_handlers.rs
@@ -1,6 +1,3 @@
 pub mod generic_request_handler;
-pub mod object_mappings;
-pub mod piece_announcement;
-pub mod piece_by_key;
-pub mod pieces_by_range;
+pub mod piece_by_index;
 pub mod segment_header;
