+++ crates/subspace-networking/src/request_handlers/piece_by_index.rs
